+++ src/index.ts
@@ -37,6 +37,12 @@ const errorCorrelationId = (error: err.Error) => {
           url: error.error.url,
           position: error.error.position,
         }
+      case 'CIRCULAR_REFERENCE': {
+        return {
+          code: error.code,
+          url: error.jsonUrl,
+        }
+      }
     }
   }
 
@@ -104,6 +110,18 @@ type Ref = {
   readonly pointer: string
 }
 
+type Specification = {
+  /**
+   * Path of `specs` JSON file
+   */
+  readonly path: string
+
+  /**
+   * readme referenced
+   */
+  readonly readMePath: string
+}
+
 const parseRef = (ref: string): Ref => {
   const i = ref.indexOf('#')
   return i < 0 ? { url: ref, pointer: '' } : { url: ref.substr(0, i), pointer: ref.substr(i + 1) }
@@ -117,53 +135,78 @@ const getReferencedFileNames = (fileName: string, doc: json.Json) => {
     .map(u => path.resolve(path.join(dir, u)))
 }
 
+const moveTo = (a: Set<string>, b: Set<string>, key: string): string => {
+  b.add(key)
+  a.delete(key)
+  return key
+}
+
 /**
- * The function finds all referenced files and put them in the `fileNames` set.
+ * The function will validate file reference as a directed graph and will detect circular reference.
+ * Detect circular reference in a directed graph using colors.
  *
- * @param readMePath a path to `readme.md` as the original source of `fileNames` set.
- * @param fileNames a set of file names from `readme.md` file.
+ * + WHITE: Vertex is not precessed yet. Initially all files mentioned in 'readme.md' is in `whiteSet`.
+ * + GRAY:  Vertex is being processed (DFS for this vertex has started, but not finished which means that
+ * all descendants (ind DFS tree) of this vertex are not processed yet (or this vertex is in function call stack)
+ * + BLACK: Vertex and all its descendants are processed
+ *
+ * For more detail: https://www.geeksforgeeks.org/detect-cycle-direct-graph-using-colors/
+ *
+ * @param current current file path
+ * @param graySet files currently being explored
+ * @param blackSet files have been explored
  */
-const resolveFileReferences = (readMePath: string, fileNames: Set<string>) =>
+const DFSTraversalValidate = (
+  current: Specification,
+  graySet: Set<string>,
+  blackSet: Set<string>,
+): asyncIt.AsyncIterableEx<err.Error> =>
   asyncIt.iterable<err.Error>(async function*() {
-    // tslint:disable-next-line:no-let
-    let fileNamesToCheck = it.toArray(fileNames)
-    // read references from `fileNamesToCheck` until there are no files are left.
-    while (fileNamesToCheck.length !== 0) {
-      // tslint:disable-next-line:readonly-array
-      const newFileNames = []
-      for (const fileName of fileNamesToCheck) {
-        // tslint:disable-next-line:no-let
-        let file: Buffer
-        // tslint:disable-next-line:no-try
-        try {
-          file = await fs.readFile(fileName)
-        } catch (e) {
-          yield {
-            code: 'NO_JSON_FILE_FOUND',
-            message: 'The JSON file is not found but it is referenced from the readme file.',
-            readMeUrl: readMePath,
-            jsonUrl: fileName,
-          }
-          continue
-        }
-        const { errors, document } = jsonParse(fileName, file.toString())
-        yield* errors
-        const refFileNames = getReferencedFileNames(fileName, document)
-        for (const refFileName of refFileNames) {
-          if (!fileNames.has(refFileName)) {
-            fileNames.add(refFileName)
-            newFileNames.push(refFileName)
-          }
-        }
-      }
-      fileNamesToCheck = newFileNames
+    if (!blackSet.has(current.path)) {
+      graySet.add(current.path)
     }
+    // tslint:disable-next-line:no-let
+    let file
+
+    // tslint:disable-next-line:no-try
+    try {
+      file = await fs.readFile(current.path)
+    } catch (e) {
+      yield {
+        code: 'NO_JSON_FILE_FOUND',
+        message: 'The JSON file is not found but it is referenced from the readme file.',
+        readMeUrl: current.readMePath,
+        jsonUrl: current.path,
+      }
+      return
+    }
+    const { errors, document } = jsonParse(current.path, file.toString())
+    yield* errors
+    const refFileNames = getReferencedFileNames(current.path, document)
+    for (const refFileName of refFileNames) {
+      if (graySet.has(refFileName)) {
+        yield {
+          code: 'CIRCULAR_REFERENCE',
+          message: `The JSON exist circular reference`,
+          readMeUrl: current.readMePath,
+          jsonUrl: current.path,
+        }
+        moveTo(graySet, blackSet, refFileName)
+      }
+
+      if (!blackSet.has(refFileName)) {
+        yield* DFSTraversalValidate({ path: refFileName, readMePath: current.readMePath }, graySet, blackSet)
+      }
+    }
+    moveTo(graySet, blackSet, current.path)
   })
 
+/**
+ * validate given `readme.md` format
+ */
 const validateReadMeFile = (readMePath: string): asyncIt.AsyncIterableEx<err.Error> =>
   asyncIt.iterable<err.Error>(async function*() {
     const file = await fs.readFile(readMePath)
-    // parse the `readme.md` file
     const m = md.parse(file.toString())
     if (!isAutoRestMd(m)) {
       yield {
@@ -175,35 +218,94 @@ const validateReadMeFile = (readMePath: string): asyncIt.AsyncIterableEx<err.Err
           'http://azure.github.io/autorest/user/literate-file-formats/configuration.html#the-file-format',
       }
     }
-    const dir = path.dirname(readMePath)
-    // get all input files from the `readme.md`.
-    const inputFiles = openApiMd.getInputFiles(m.markDown).toArray()
-    // normalize the file names.
-    const inputFileSet = inputFiles
-      .map(f => path.resolve(path.join(dir, ...f.split('\\'))))
-      .reduce((s, v) => s.add(v), new Set<string>())
-    // add all referenced files to the `set`
-    yield* resolveFileReferences(readMePath, inputFileSet)
-    // report errors if the `dir` folder has JSON files which are not referenced.
-    yield* fs
-      .recursiveReaddir(dir)
-      .filter(filePath => path.extname(filePath) === '.json' && !inputFileSet.has(filePath))
-      .map<err.Error>(filePath => ({
+  })
+
+/**
+ * Validate spec files in two steps:
+ * 1. `DFSTraversalValidate`: Analyze specs as a directed graph to detect circular reference and
+ *     generate `blackSet` that contains all explored specs.
+ * 2.  Get difference set between `allInputFileSet` and `blackSet`, and then report `UNREFERENCED_JSON_FILE` error.
+ *
+ * @param inputFileSet files referenced from 'readme.md' is the subset of `allInputFileSet`
+ * @param allInputFileSet files appear in specification folder.
+ */
+const validateInputFiles = (
+  inputFileSet: Set<Specification>,
+  allInputFileSet: Set<Specification>,
+): asyncIt.AsyncIterableEx<err.Error> =>
+  // tslint:disable-next-line: no-async-without-await
+  asyncIt.iterable<err.Error>(async function*() {
+    // report errors if the `dir` folder has JSON files where exist circular reference
+    const graySet = new Set<string>()
+    const blackSet = new Set<string>()
+    for (const current of inputFileSet) {
+      yield* DFSTraversalValidate(current, graySet, blackSet)
+    }
+
+    // report errors if the `dir` folder has JSON files which are not referenced
+    yield* asyncIt
+      .fromSync(allInputFileSet.values())
+      .filter(spec => !blackSet.has(spec.path))
+      .map<err.Error>(spec => ({
         code: 'UNREFERENCED_JSON_FILE',
         message: 'The JSON file is not referenced from the readme file.',
-        readMeUrl: readMePath,
-        jsonUrl: path.resolve(filePath),
+        readMeUrl: spec.readMePath,
+        jsonUrl: spec.path,
       }))
   })
 
+const getInputFilesFromReadme = (readMePath: string): asyncIt.AsyncIterableEx<Specification> =>
+  asyncIt.iterable<Specification>(async function*() {
+    const file = await fs.readFile(readMePath)
+    const m = md.parse(file.toString())
+    const dir = path.dirname(readMePath)
+
+    yield* openApiMd
+      .getInputFiles(m.markDown)
+      .uniq()
+      .map(f => path.resolve(path.join(dir, ...f.split('\\'))))
+      .map<Specification>(f => ({ path: f, readMePath }))
+  })
+
+const getAllInputFilesUnderReadme = (readMePath: string): asyncIt.AsyncIterableEx<Specification> =>
+  // tslint:disable-next-line: no-async-without-await
+  asyncIt.iterable<Specification>(async function*() {
+    const dir = path.dirname(readMePath)
+    yield* fs
+      .recursiveReaddir(dir)
+      .filter(filePath => path.extname(filePath) === '.json')
+      .map<Specification>(filePath => ({ path: filePath, readMePath }))
+  })
+
+/**
+ * Validate global specification folder and prepare arguments for `validateInputFiles`.
+ */
 const validateSpecificationFolder = (cwd: string) =>
   asyncIt.iterable<err.Error>(async function*() {
     const specification = path.resolve(path.join(cwd, 'specification'))
+
     if (await fs.exists(specification)) {
-      yield* fs
+      const allReadMeFiles = fs
         .recursiveReaddir(specification)
         .filter(f => path.basename(f).toLowerCase() === 'readme.md')
-        .flatMap(validateReadMeFile)
+
+      yield* allReadMeFiles.flatMap(validateReadMeFile)
+
+      const referencedFiles = await allReadMeFiles
+        .flatMap(getInputFilesFromReadme)
+        .fold((fileSet: Set<Specification>, spec) => {
+          fileSet.add(spec)
+          return fileSet
+        }, new Set<Specification>())
+
+      const allFiles = await allReadMeFiles
+        .flatMap(getAllInputFilesUnderReadme)
+        .fold((fileSet: Set<Specification>, spec) => {
+          fileSet.add(spec)
+          return fileSet
+        }, new Set<Specification>())
+
+      yield* validateInputFiles(referencedFiles, allFiles)
     }
   })
 
@@ -215,7 +317,6 @@ const avocadoForDir = async (cwd: string) => {
   for await (const e of validateSpecificationFolder(cwd)) {
     map.set(errorCorrelationId(e), e)
   }
-
   return map
 }